Updates for Wiki Drag and Drop
